Some experimentation with glue code for the siox-ll-api and support for mixed C/C++. Ontology offers a in-memory storage for datatypes but does not provide IDs so far.
The basic idea is to implement the C interfaces in monitoring/low-level-c/siox-ll.cpp
By providing glue code to C++ objects and implementations it will load the configuration and instantiate
the implementations for Ontology and the Monitoring path.
This seems to be a good idea but the detailed interfaces have to be worked out.